satellite/admin-ui: add delete project UI
This change adds UI functionality for deleting projects. If abbreviated
project deletion is enabled, an action is available for marking projects
as pending deletion without requiring that the project is empty. Because
of this, project status can not be set to pending deletion via the
project update dialog.

Issue: https://github.com/storj/storj-private/issues/1634

Change-Id: Ia1dd0aff9587651a8d4be9179b8e20e072aed88e