storagenode/hashstore: full table reconstruction
log reconciliation is basically table
reconstruction in the case where every log file
is included, so we can just do it automatically
with very little extra code.

we don't want to always do it because it can
hide problems that happened, so we have a config
boolean that is default off but if you enable it
will cause a missing hashtbl to reconcile all
logs during startup.

Change-Id: I2b943d96c881eae7c6b1485a3e77efd89192c2ac