satellite/admin: allow abbreviated project deletion
This change makes it so that when abbreviated project deletion is on the
satellite UI, projects can be marked PendingDeletion to be cleaned up by
a chore later even if they still contain buckets and objects. For this
reason, the update project endpoint will forbid setting status to
PendingDeletion. This status can only be set through the disable project
endpoint.

Issue: https://github.com/storj/storj-private/issues/1634

Change-Id: I0a63bd1cc0449e30785f4a5fdd559ebb319afcc3