satellite/admin: allow abbreviated user deletion
This change makes it so that when abbreviated user deletion is enabled
on the satellite UI, users can be marked PendingDeletion to be cleaned
up by a chore later even if they still have projects and data. For this
reason, the update user endpoint will forbid setting status to
PendingDeletion. This status can only be set via the disable user
endpoint.
It also removes the hasUnpaidInvoices field from the user struct.

Issue: https://github.com/storj/storj-private/issues/1635

Change-Id: Ibfa2c5ee463f28ce343f79e2348ad809360694d4