satellite/admin-ui: allow abbreviated user deletion
This change makes adds a new account action "Mark Pending Deletion" this
action will be available if abbreviated user deletion is enabled on the
satellite. It will allow the admin to mark a user for deletion without
requiring that the user has no active projects.
Because of this, user status cannot be set to "Pending Deletion" via the
account update dialog anymore, it must be done via the new action.

Issue: https://github.com/storj/storj-private/issues/1635

Change-Id: I0ffd28469931b53ebe3a7e0b5f4d194c70b6dbfb