satellite/console: add config for partner UI mods
This change adds a new config for partner UI modifications that were
previously stored in the tardigrade-satellite-theme repository.
The config can be provided as a YAML string or a path to a YAML file.
An endpoint is added to retrieve these configs for use in the front-end.

Issue: https://github.com/storj/storj-private/issues/1629

Change-Id: Iae75575de4890e71d6d0e121ab714de3c4e507d1