satellite/admin-ui: fix user/project state update bug
This change fixes a bug where after a user/project is update from a
dialog, the user/project detail page will vanish and appear again with
the dialog still open. This happens because when the new state is being
fetched after the update, the current state is set to null. This change
removes the clearing of current user/project state before fetching the
new state.

Change-Id: Ia0ea99c00b2a01305f7ed64322fc44ddb513da95