storagenode: Fix UsedSpace calculation of Storagenode web console
Two problems:

 * Shows 0 usage when Using the (experimental) dedicated disk space reporter
 * Shows a confusing used space when hashstore is used (binary overhead / dead bytes are not included)

The root cause is the same: we used "UsedByPiece" instead of "AllUsedPiece"

The solution is to report the full "UsedSpace" together with more granular data (UsedByTrash, ...)

Unfortunatelly this makes the web ui complex: I wouldn't like to add more than 3-4 lines of legend.

As a solution Legend is simplified (Trash is not displayed). But the `Used` part of the Doughnut is split, and Reclaimable and Trash are displayed there.

Change-Id: I2bc2ccbb375aac3a0ceaa411d90155dcad82383f