satellite/{console,web}: prevent abbreviated project deletion if object lock is enabled
This change ensures that if a project has buckets with Object Lock
enabled, the user will not be able to mark the project for deletion via
abbreviated project deletion.
The pending deletion cleanup chore is also updated to skip projects with
lock-enabled buckets.

Issue: https://github.com/storj/storj-private/issues/1657

Change-Id: Id811033d97324ab1a1c7249b9ed8764e34d300c2