storagenode/hashstore: windows fixes
operators reported errors going from 1.141.4 to
1.142.4 on windows and the errors indicated that
it was during atomic file rename. the two changes
during that time were the switch from os.Rename
to platform.Rename and the removal of share
permissions on CreateFile.

this change puts back the share permissions for
the CreateFile api on windows. dunno why i removed
it in the first place. it does keep the os.Rename
to platform.Rename change because it should be the
same except for the windows.MOVEFILE_WRITE_THROUGH
which should make it more safe for atomic rename
because it's documented to make it not return
until the file is actually moved on disk.

the tests pass on wine which is the best i can do
right now.

Change-Id: Ibe80fb078a928890d2badff5f99314777cabd8cc