satellite: whitelabel email system
This change updates the emailing system to support branding emails for
different tenants. This includes changing links, names, addresses and
colours in emails to match what is configured for a specific tenant. If
no tenant is detected in context, the default branding is used.

Issue: https://github.com/storj/storj-private/issues/1650

Change-Id: I2d9d1219fc5ddba5cfca132fd665fb5a8804f5f0