shared/{h/m}otel: monkit helpers to use open telemetry
Monkit, our observability tool, can handle metrics/traces, with all the existing strength and weaknesses:

It has a very easy-to-use API, and well tested. On the other hand, it coulnd't be used with modern Observability 2.0 application:

 1. the data model is incompatible (bit size of spans/traces)
 2. metrics are not connected to spans (ctx)
 3. span based events (span annotations, which work like evenkit) are not very well supported

It would be nice to test some advanced observability tools, without big changes in the codespace.

This patch is a POC, and provides two helpers to make it possible.

 1. Monkit for Opentelemetry (motel package)

   This is an Opentelemetry backend, using Monkit.

   With the following two lines of code, we can use Opentelemetry API on client side, but all spans will be managed by Monkit.

```
otel.SetMeterProvider(motel.NewMeterProvider(monkit.Default))
otel.SetTracerProvider(motel.NewTraceProvider(monkit.Default))
```

   It is a change only for the client code.

   Benefit: in the future, we can switch to use different backend (like the official otel backend, or a new custom one), and experiment with different tools.

 2. Helpers for Opentelemetry (hotel package)

   But there are cases where our Monkit API can be prefered as it's shorter (like the legendary `defer mon.Task()(&ctx)(&err)` one liner)

   To keep the code short, we can use some helpers / wrappers, which can utilize Opentelemetry under the hood, but can simplify the usage of `otel`.

   This helper / wrapper should be as close to the monkit API as possible. In this case we can easily switch to use the wrapper with updating only the import names (from Monkit to motel)

These packages are intentionally kept here and not in monkit repo: it's just partial implementation / poc. I expect way more changes, and I would like to avoid frequent bumps.

Once is stabilized, we can consider exporting it to monkit repo (if it's required).

Change-Id: I76c98f2c04cbdbc64a33c8e232b13630fca87d92