web/satellite: make UI config backward compatible
This change reintroduces the removed config files in
web/satellite/src/config. This will make it so that in the case where
the backend is not configured with the expected configurations, the
config files will be used as fallback. The content of the config files
will be provided by the tardigrade-satellite repository.

Change-Id: I5904c9ac493f7c890a59973355a28677060b094a